Rename SickNoteAction to SickNoteCommentAction
