+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/SickNoteCommentAction.java
@@ -3,7 +3,7 @@ package org.synyx.urlaubsverwaltung.sicknote;
 /**
  * Executed action on a sick note.
  */
-public enum SickNoteAction {
+public enum SickNoteCommentAction {
 
     CREATED,
     EDITED,
